cleaned code of old code and comments, needs more formatting, added toast
