+++ SHPE-UF-Mobile-Swift/SHPE-UF-Mobile-Swift/Assets.xcassets/swift.gradcap.imageset/Contents.json
