+++ SHPE-UF-Mobile-Swift/SHPE-UF-Mobile-Swift/SHPEUFAppView.swift
@@ -1,16 +1,14 @@
-//
-//  ContentView.swift
-//  SHPE-UF-Mobile-Swift
-//
-//  Created by Jesus Lopez on 10/19/23.
-//
+
 
 import SwiftUI
 
 struct SHPEUFAppView: View {
     let requestHandler = RequestHandler()
-    var body: some View {
+    var body: some View 
+    {
         RegisterView()
+       // AcademicView(viewModel: RegisterViewModel())
+        
             .onAppear {
                 print("I HAVE APPEARED")
             }
