--- SHPE-UF-Mobile-Swift/SHPE-UF-Mobile-Swift/ui/pages/register/SHPEITO.swift
@@ -1,8 +1,0 @@
-//
-//  SHPEITO.swift
-//  SHPE-UF-Mobile-Swift
-//
-//  Created by test on 2/1/24.
-//
-
-import Foundation
