+++ SHPE-UF-Mobile-Swift/SHPE-UF-Mobile-Swift/ui/pages/register/placeholderForLandingPageView.swift
@@ -1,0 +1,18 @@
+//
+//  placeholderForLandingPageView.swift
+//  SHPE-UF-Mobile-Swift
+//
+//  Created by test on 2/29/24.
+//
+
+import SwiftUI
+
+struct placeholderForLandingPageView: View
+{
+    @Environment(\.presentationMode) var isPresented
+    @StateObject var viewModel: RegisterViewModel
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
